Coerce Bernoulli probability tables to the data.frame

Auxiliary function for "bernoulli_naive_bayes" objects for extracting and coercing the Bernoulli probability tables to the data.frame.

bernoulli_tables_to_df(object, format = "wide")

Arguments

object
object of class inheriting from "bernoulli_naive_bayes".

format
character string with the format of the resulting data.frame - either "wide"" or "long".

Value

data.frame with class conditional Bernoulli probabilities.

Details

This auxiliary function coerces the Bernoulli probability tables to a data frame with either wide or long format.

See also

Examples

# Simulate data
cols <- 10 ; rows <- 100 ; probs <- c("0" = 0.4, "1" = 0.1)
M <- matrix(sample(0:1, rows * cols,  TRUE, probs), nrow = rows, ncol = cols)
y <- factor(sample(paste0("class", LETTERS[1:2]), rows, TRUE, prob = c(0.3,0.7)))
colnames(M) <- paste0("V", seq_len(ncol(M)))
laplace <- 0.5

# Fit the Bernoulli Naive Bayes model
bnb <- bernoulli_naive_bayes(x = M, y = y, laplace = laplace)

# Extract the probability tables and convert them to the data.frame
bernoulli_tables_to_df(bnb)

#>      classA:0  classA:1  classB:0  classB:1
#> V1  0.8166667 0.1833333 0.7986111 0.2013889
#> V2  0.7833333 0.2166667 0.8541667 0.1458333
#> V3  0.8833333 0.1166667 0.8541667 0.1458333
#> V4  0.7833333 0.2166667 0.8263889 0.1736111
#> V5  0.8166667 0.1833333 0.7569444 0.2430556
#> V6  0.7500000 0.2500000 0.7986111 0.2013889
#> V7  0.7833333 0.2166667 0.8402778 0.1597222
#> V8  0.7166667 0.2833333 0.7152778 0.2847222
#> V9  0.6166667 0.3833333 0.8680556 0.1319444
#> V10 0.9500000 0.0500000 0.7986111 0.2013889

# Compare first row with the first table:
tables(bnb, which = 1)

#> 
#> --------------------------------------------------------------------------------- 
#>  ::: V1 (Bernoulli) 
#> --------------------------------------------------------------------------------- 
#>      classA    classB
#> 0 0.8166667 0.7986111
#> 1 0.1833333 0.2013889
#> 
#> ---------------------------------------------------------------------------------