Multinomial Naive Bayes Classifier

multinomial_naive_bayes is used to fit the Multinomial Naive Bayes model.

multinomial_naive_bayes(x, y, prior = NULL, laplace = 0.5, ...)

Arguments

x
numeric matrix with integer predictors (matrix or dgCMatrix from Matrix package).

y
class vector (character/factor/logical).

prior
vector with prior probabilities of the classes. If unspecified, the class proportions for the training set are used. If present, the probabilities should be specified in the order of the factor levels.

laplace
value used for Laplace smoothing (additive smoothing). Defaults to 0.5.

...
not used.

Value

multinomial_naive_bayes returns an object of class "multinomial_naive_bayes" which is a list with following components:

data

list with two components: x (matrix with predictors) and y (class variable).

levels

character vector with values of the class variable.

laplace

amount of Laplace smoothing (additive smoothing).

params

matrix with class conditional parameter estimates.

prior

numeric vector with prior probabilities.

call

the call that produced this object.

Details

This is a specialized version of the Naive Bayes classifier, where the features represent frequencies generated by a multinomial distribution.

Sparse matrices of class "dgCMatrix" (Matrix package) are supported in order to speed up calculation times.

Please note that the Multinomial Naive Bayes is not available through the naive_bayes function.

References

McCallum, Andrew; Nigam, Kamal (1998). A comparison of event models for Naive Bayes text classification (PDF). AAAI-98 workshop on learning for text categorization. 752. http://www.cs.cmu.edu/~knigam/papers/multinomial-aaaiws98.pdf

See also

Examples

# library(naivebayes)

### Simulate the data:
cols <- 10 ; rows <- 100
M <- matrix(sample(0:5, rows * cols, TRUE, prob = c(0.95, rep(0.01, 5))), nrow = rows, ncol = cols)
y <- factor(sample(paste0("class", LETTERS[1:2]), rows, TRUE, prob = c(0.3,0.7)))
colnames(M) <- paste0("V", seq_len(ncol(M)))
laplace <- 1

### Train the Multinomial Naive Bayes
mnb <- multinomial_naive_bayes(x = M, y = y, laplace = laplace)
summary(mnb)

#> 
#> ============================ Multinomial Naive Bayes ============================ 
#>  
#> - Call: multinomial_naive_bayes(x = M, y = y, laplace = laplace) 
#> - Laplace: 1 
#> - Classes: 2 
#> - Samples: 100 
#> - Features: 10 
#> - Prior probabilities: 
#>     - classA: 0.31
#>     - classB: 0.69
#> 
#> --------------------------------------------------------------------------------- 

# Classification
head(predict(mnb, newdata = M, type = "class")) # head(mnb %class% M)

#> [1] classB classB classA classB classB classB
#> Levels: classA classB

# Posterior probabilities
head(predict(mnb, newdata = M, type = "prob")) # head(mnb %prob% M)

#>           classA      classB
#> [1,] 0.400244325 0.599755675
#> [2,] 0.310000000 0.690000000
#> [3,] 0.990910425 0.009089575
#> [4,] 0.310000000 0.690000000
#> [5,] 0.003904309 0.996095691
#> [6,] 0.310000000 0.690000000

# Parameter estimates
coef(mnb)

#>         classA     classB
#> V1  0.08196721 0.06896552
#> V2  0.01639344 0.19540230
#> V3  0.18032787 0.08045977
#> V4  0.21311475 0.10344828
#> V5  0.09836066 0.06896552
#> V6  0.13114754 0.05747126
#> V7  0.01639344 0.11494253
#> V8  0.13114754 0.11494253
#> V9  0.08196721 0.03448276
#> V10 0.04918033 0.16091954

### Sparse data: train the Multinomial Naive Bayes
library(Matrix)
M_sparse <- Matrix(M, sparse = TRUE)
class(M_sparse) # dgCMatrix

#> [1] "dgCMatrix"
#> attr(,"package")
#> [1] "Matrix"

# Fit the model with sparse data
mnb_sparse <- multinomial_naive_bayes(M_sparse, y, laplace = laplace)

# Classification
head(predict(mnb_sparse, newdata = M_sparse, type = "class"))

#> [1] classB classB classA classB classB classB
#> Levels: classA classB

# Posterior probabilities
head(predict(mnb_sparse, newdata = M_sparse, type = "prob"))

#>           classA      classB
#> [1,] 0.400244325 0.599755675
#> [2,] 0.310000000 0.690000000
#> [3,] 0.990910425 0.009089575
#> [4,] 0.310000000 0.690000000
#> [5,] 0.003904309 0.996095691
#> [6,] 0.310000000 0.690000000

# Parameter estimates
coef(mnb_sparse)

#>         classA     classB
#> V1  0.08196721 0.06896552
#> V2  0.01639344 0.19540230
#> V3  0.18032787 0.08045977
#> V4  0.21311475 0.10344828
#> V5  0.09836066 0.06896552
#> V6  0.13114754 0.05747126
#> V7  0.01639344 0.11494253
#> V8  0.13114754 0.11494253
#> V9  0.08196721 0.03448276
#> V10 0.04918033 0.16091954