Multinomial Naive Bayes Classifier

multinomial_naive_bayes is used to fit the Multinomial Naive Bayes model.

multinomial_naive_bayes(x, y, prior = NULL, laplace = 0.5, ...)

Arguments

x
numeric matrix with integer predictors.

y
class vector (character/factor/logical).

prior
vector with prior probabilities of the classes. If unspecified, the class proportions for the training set are used. If present, the probabilities should be specified in the order of the factor levels.

laplace
value used for Laplace smoothing (additive smoothing). Defaults to 0.5.

...
not used.

Value

multinomial_naive_bayes returns an object of class "multinomial_naive_bayes" which is a list with following components:

data

list with two components: x (matrix with predictors) and y (class variable).

levels

character vector with values of the class variable.

laplace

amount of Laplace smoothing (additive smoothing).

params

matrix with class conditional parameter estimates.

prior

numeric vector with prior probabilities.

call

the call that produced this object.

Details

This is a specialized version of the Naive Bayes classifier, where the features represent the frequencies with which events have been generated by a multinomial distribution.

The Multinomial Naive Bayes is not available through the naive_bayes function.

References

McCallum, Andrew; Nigam, Kamal (1998). A comparison of event models for Naive Bayes text classification (PDF). AAAI-98 workshop on learning for text categorization. 752. http://www.cs.cmu.edu/~knigam/papers/multinomial-aaaiws98.pdf

See also

Examples

### Simulate the data:
cols <- 10 ; rows <- 100
M <- matrix(sample(0:5, rows * cols,  TRUE), nrow = rows, ncol = cols)
y <- factor(sample(paste0("class", LETTERS[1:2]), rows, TRUE, prob = c(0.3,0.7)))
colnames(M) <- paste0("V", seq_len(ncol(M)))
laplace <- 1

### Train the Multinomial Naive Bayes
mnb <- multinomial_naive_bayes(x = M, y = y, laplace = laplace)
summary(mnb)

#> 
#> ============================ Multinomial Naive Bayes ============================ 
#>  
#> - Call: multinomial_naive_bayes(x = M, y = y, laplace = laplace) 
#> - Laplace: 1 
#> - Classes: 2 
#> - Samples: 100 
#> - Features: 10 
#> - Prior probabilities: 
#>     - classA: 0.29
#>     - classB: 0.71
#> 
#> --------------------------------------------------------------------------------- 

# Classification
head(predict(mnb, newdata = M, type = "class"))

#> [1] classB classB classB classB classB classB
#> Levels: classA classB

# Posterior probabilities
head(predict(mnb, newdata = M, type = "prob"))

#>         classA    classB
#> [1,] 0.1967351 0.8032649
#> [2,] 0.3540362 0.6459638
#> [3,] 0.3896224 0.6103776
#> [4,] 0.1204691 0.8795309
#> [5,] 0.1879130 0.8120870
#> [6,] 0.3381093 0.6618907

# Parameter estimates
coef(mnb)

#>         classA     classB
#> V1  0.09517426 0.10374640
#> V2  0.09249330 0.09279539
#> V3  0.09785523 0.08645533
#> V4  0.08445040 0.10432277
#> V5  0.10589812 0.11585014
#> V6  0.10187668 0.09855908
#> V7  0.10455764 0.08703170
#> V8  0.10723861 0.11700288
#> V9  0.11796247 0.10144092
#> V10 0.09249330 0.09279539