Predict Method for multinomial_naive_bayes Objects

Classification based on the Multinomial Naive Bayes model.

# S3 method for multinomial_naive_bayes
predict(object, newdata = NULL, type = c("class","prob"), ...)

Arguments

object
object of class inheriting from "multinomial_naive_bayes".

newdata
matrix with non-negative integer predictors (only numeric matrix is accepted).

type
if "class", new data points are classified according to the highest posterior probabilities. If "prob", the posterior probabilities for each class are returned.

...
not used.

Value

predict.multinomial_naive_bayes returns either a factor with class labels corresponding to the maximal conditional posterior probabilities or a matrix with class label specific conditional posterior probabilities.

Details

This is a specialized version of the Naive Bayes classifier, where the features represent the frequencies with which events have been generated by a multinomial distribution.

The Multinomial Naive Bayes is not available through the naive_bayes function.

The NAs in the newdata are not included into the calculation of posterior probabilities; and if present an informative warning is given.

References

McCallum, Andrew; Nigam, Kamal (1998). A comparison of event models for Naive Bayes text classification (PDF). AAAI-98 workshop on learning for text categorization. 752. http://www.cs.cmu.edu/~knigam/papers/multinomial-aaaiws98.pdf

See also

Examples

### Simulate the data:
cols <- 10 ; rows <- 100
M <- matrix(sample(0:5, rows * cols,  TRUE), nrow = rows, ncol = cols)
y <- factor(sample(paste0("class", LETTERS[1:2]), rows, TRUE, prob = c(0.3,0.7)))
colnames(M) <- paste0("V", seq_len(ncol(M)))
laplace <- 1

### Train the Multinomial Naive Bayes
mnb <- multinomial_naive_bayes(x = M, y = y, laplace = laplace)
mnb

#> 
#> ============================ Multinomial Naive Bayes ============================ 
#>  
#>  Call: 
#> multinomial_naive_bayes(x = M, y = y, laplace = laplace)
#> 
#> --------------------------------------------------------------------------------- 
#>  
#> Laplace smoothing: 1
#> 
#> --------------------------------------------------------------------------------- 
#>  
#>  A priori probabilities: 
#> classA classB 
#>   0.28   0.72 
#> 
#> --------------------------------------------------------------------------------- 
#>  
#>         Classes
#> Features     classA     classB
#>      V1  0.10252809 0.10311493
#>      V2  0.10955056 0.09989259
#>      V3  0.11516854 0.10902256
#>      V4  0.09831461 0.09935553
#>      V5  0.09831461 0.10365199
#>      V6  0.11516854 0.10955961
#>      V7  0.09129213 0.09774436
#>      V8  0.09129213 0.08915145
#>      V9  0.08286517 0.09076262
#>      V10 0.09550562 0.09774436
#> 
#> --------------------------------------------------------------------------------- 
#> 
#> 

# Classification
head(predict(mnb, newdata = M, type = "class"))

#> [1] classB classB classB classB classB classB
#> Levels: classA classB

# Posterior probabilities
head(predict(mnb, newdata = M, type = "prob"))

#>         classA    classB
#> [1,] 0.1941549 0.8058451
#> [2,] 0.3030244 0.6969756
#> [3,] 0.2360529 0.7639471
#> [4,] 0.2919285 0.7080715
#> [5,] 0.2379393 0.7620607
#> [6,] 0.1958530 0.8041470