Predict Method for naive_bayes Objects

Classification based on Naive Bayes models.

# S3 method for naive_bayes
predict(object, newdata = NULL, type = c("class","prob"),
  threshold = 0.001, eps = 0, ...)

Arguments

object
object of class inheriting from "naive_bayes".

newdata
matrix or dataframe with categorical (character/factor/logical) or metric (numeric) predictors.

type
if "class", new data points are classified according to the highest posterior probabilities. If "prob", the posterior probabilities for each class are returned.

threshold
value by which zero probabilities or probabilities within the epsilon-range corresponding to metric variables are replaced (zero probabilities corresponding to categorical variables can be handled with Laplace (additive) smoothing).

eps
value that specifies an epsilon-range to replace zero or close to zero probabilities by threshold. It applies to metric variables.

...
not used.

Value

predict.naive_bayes returns either a factor with class labels corresponding to the maximal conditional posterior probabilities or a matrix with class label specific conditional posterior probabilities.

Details

Computes conditional posterior probabilities for each class label using the Bayes' rule under the assumption of independence of predictors. If no new data is provided, the data from the object is used. Logical variables are treated as categorical (binary) variables. Predictors with missing values are not included into the computation of posterior probabilities.

See also

Examples

ind_iris <- sample(1:nrow(iris), size = round(0.3 * nrow(iris)))
iris_train <- iris[-ind_iris, ]
iris_test <- iris[ind_iris, ]

nb_iris <- naive_bayes(Species ~ ., iris_train)
predict(nb_iris, iris_test)

#>  [1] setosa     setosa     virginica  virginica  setosa     virginica 
#>  [7] setosa     versicolor virginica  setosa     versicolor setosa    
#> [13] virginica  virginica  virginica  virginica  versicolor versicolor
#> [19] setosa     setosa     virginica  virginica  virginica  versicolor
#> [25] virginica  virginica  versicolor setosa     setosa     versicolor
#> [31] setosa     virginica  setosa     setosa     setosa     virginica 
#> [37] versicolor versicolor virginica  setosa     setosa     setosa    
#> [43] versicolor virginica  versicolor
#> Levels: setosa versicolor virginica

head(predict(nb_iris, iris_test, type = "prob"))

#>             setosa   versicolor    virginica
#> [1,]  1.000000e+00 4.054291e-15 2.281728e-21
#> [2,]  1.000000e+00 1.099206e-18 2.831382e-25
#> [3,] 9.389589e-134 2.929155e-02 9.707085e-01
#> [4,] 1.158962e-221 5.203778e-11 1.000000e+00
#> [5,]  1.000000e+00 4.340931e-17 3.210193e-24
#> [6,] 9.389589e-134 2.929155e-02 9.707085e-01

# NOT RUN {
vars <- 10
rows <- 500000
y <- sample(c("a", "b"), rows, TRUE)

# Only categorical variables
X1 <- as.data.frame(matrix(sample(letters[5:9], vars * rows, TRUE),
                           ncol = vars))
nb_cat <- naive_bayes(x = X1, y = y)
nb_cat
system.time(pred2 <- predict(nb_cat, X1))

# Only numeric variables
X2 <- as.data.frame(matrix(rnorm(vars * rows), ncol = vars))
nb_num <- naive_bayes(x = X2, y = y)
nb_num
system.time(pred2 <- predict(nb_num, X2))
# }